Fixed difficult to read text elements
These changes were made to fix readability issues with very light
text on a light background in the Update Packages dialog, as well
as dark text on the dark background of the Options dialog.

Comments were also cleaned up to make the purpose of SCSS elements
easier to understand at first glance.